+++ crates/affinidi-tdk/common/affinidi-data-integrity/examples/sign_and_verify.rs
@@ -1,11 +1,11 @@
-use affinidi_data_integrity::{DataIntegrityProof, verification_proof::verify_data};
-use affinidi_did_resolver_cache_sdk::{DIDCacheClient, config::DIDCacheConfigBuilder};
+use affinidi_data_integrity::{
+    DataIntegrityProof, verification_proof::verify_data_with_public_key,
+};
 use affinidi_secrets_resolver::secrets::Secret;
 use serde_json::json;
 use tracing_subscriber::filter;
 
-#[tokio::main]
-async fn main() {
+fn main() {
     // construct a subscriber that prints formatted traces to stdout
     let subscriber = tracing_subscriber::fmt()
         // Use a more compact, abbreviated log format
@@ -68,11 +68,7 @@ async fn main() {
     let proof = DataIntegrityProof::sign_jcs_data(&input_doc, None, &secret, None)
         .expect("Couldn't sign Document");
 
-    let resolver = DIDCacheClient::new(DIDCacheConfigBuilder::default().build())
-        .await
-        .unwrap();
-    let _ = verify_data(&resolver, &input_doc, None, &proof)
-        .await
+    let _ = verify_data_with_public_key(&input_doc, None, &proof, secret.get_public_bytes())
         .expect("Couldn't validate doc");
 
     println!(